% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{build_qc_report}
\alias{build_qc_report}
\title{生成 HTML 数据质控报告}
\usage{
build_qc_report(x, path)
}
\arguments{
\item{x}{\code{\link[=run_qc]{run_qc()}} 产出的结果列表}

\item{path}{HTML 输出路径}
}
\description{
生成 HTML 数据质控报告
}
