% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale.R
\name{center_scale}
\alias{center_scale}
\title{RClean center and scale}
\usage{
center_scale(indices, DF, center, scale)
}
\arguments{
\item{indices}{the indices from the shiny gadget}

\item{DF}{the dataframe that is being worked on}

\item{center}{binary true/false for scale() function}

\item{scale}{binary true/false for scale() function}
}
\description{
center and scale function to handle non-numeric data
}
