% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean --- Clean code.}
\usage{
clean(file, var, libs, format)
}
\arguments{
\item{file}{File path to a script.}

\item{var}{The name of a variable of interest.}

\item{libs}{LOGICAL: should library dependencies be detected and
included?}

\item{reformat}{LOGICAL: should the minimized code be re-formatted
following common syntax?}

\item{plot}{LOGICAL: should a network diagram of the cleaned script
be produced?}

\item{rp}{LOGICAL: should retrospective provenance be used? This
should be made available in JSON format as an object within
options (i.e. options(prov.json = <your retrospective
provenance object>).}
}
\value{
A character vector containing a minimized script based on
    the given input variable.
}
\description{
Reduce the size and complexity of a analytical script through
algorithmic detection of the essential code needed to produce a set
of results.
}
\author{
Matthew K. Lau
}
