% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanR.R
\name{cleanR}
\alias{cleanR}
\title{cleanR --- Produce "clean" code.
OUTPUT = A workspace containing a cleanced script and associated
data files.}
\usage{
cleanR(file = "Path to an R script", result, save.comments = FALSE)
}
\arguments{
\item{file}{Path to an R script.}

\item{result}{A desired output present in the script.}

\item{save.comments}{LOGICAL: should comments be preserved in cleaned code?}
}
\value{
Workspace with the minimal code needed to produce the
specified results.
}
\description{
Produces simplifed, "cleaned" code that generates
a set of output from an initial, "draft" script.
}
\author{
Matthew K. Lau
}
