% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_graph.R
\name{code_graph}
\alias{code_graph}
\title{code_graph --- A network diagram of computational dependencies.}
\usage{
code_graph(script, pdf_file)
}
\arguments{
\item{script}{Script object from CodeDepends or a file path to a script.}

\item{pdf_file}{Path to output PDF of the code graph.}
}
\value{
A network plot.
}
\description{
Plot the dependencies among functions and objects for a given script.
}
\examples{
\dontrun{
script <- system.file("example", "simple_script.R", package = "Rclean")
simple.script <- readScript(script)
code_graph(simple.script)
}
}
\author{
Matthew K. Lau
}
