% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeGraph.R
\name{codeGraph}
\alias{codeGraph}
\title{codeGraph --- A network diagram of computational dependencies.}
\usage{
codeGraph(g)
}
\arguments{
\item{g}{Matrix representation of the graph.}
}
\value{
A network plot.
}
\description{
Plot the dependencies among functions and objects for a given script.
}
\author{
Matthew K. Lau
}
