% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.libs.R
\name{get.libs}
\alias{get.libs}
\title{get.libs --- Determine the libraries needed for a given script.}
\usage{
get.libs(src)
}
\arguments{
\item{src}{Character vector containing the source script.}
}
\value{
The libraries used for each step of the script.
}
\description{
Uses code dependency information to produce a set of necessary libraries.
}
\author{
Matthew K. Lau
}
