% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libs.R
\name{get_libs}
\alias{get_libs}
\title{get_libs --- Determine the libraries needed for a given script.}
\usage{
get_libs(src)
}
\arguments{
\item{src}{}
}
\value{
The libraries used for each step of the script.
}
\description{
Uses code dependency information to produce a set of necessary
libraries.
}
\author{
Matthew K. Lau
}
