% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{get_vars --- Get the variables that are created in a script.}
\usage{
get_vars(script)
}
\arguments{
\item{script}{File path to a script or a Script object from
CodeDepends.}
}
\value{
A list of variables from a given script.
}
\description{
Determining the variables that are created in a script is useful
for orienting code cleaning (i.e. refactoring) efforts.
}
\examples{
script <- system.file(
        "example", 
        "simple_script.R", 
        package = "Rclean")
get_vars(script)
}
\author{
Matthew K. Lau
}
