% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Firsher's Iris dataset}
\format{A \code{tibble} with 200 rows and 3 variables

\describe{
  \item{PSC}{Contract identifier for each product and service}
  \item{x_attribute}{x attribute score, from 1 (worst) to 5 (best) in increments of 0.01}
  \item{y_attribute}{y attribute score, from 1 (worst) to 10 (best) in increments of 0.01}
}}
\usage{
iris
}
\description{
A dataset containing a . Dataset taken from kaggle, in .xls format, from 
https://www.kaggle.com/saurabh00007/iriscsv
}
\keyword{datasets}
