% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{keep --- Copy code to clipboard or write a specified file.
OUTPUT = Copies code to the clipboard or writes out code from an
object to a specified file.}
\usage{
keep(x, file = NULL)
}
\arguments{
\item{x}{Line(s) of code.}

\item{file}{Path to write code. If not specified, code is copied to
the clipboard.}
}
\value{
The code is copied to the clipboard or written to disk.
}
\description{
Simplifies writing and copying code to clipboard by removing the
need to separately create and specify a file connection or to send
code to the clipboard.
}
\examples{
\dontrun{
script <- system.file(
        "example", 
        "simple_script.R", 
        package = "Rclean")
clean.code <- clean(script, “tab.15”)
## Copies code to your clipboard
keep(clean.code)
## Saves code to your disk
keep(clean.code, file = "clean_code.R")
}
}
\author{
Matthew K. Lau
}
