% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.code.R
\name{keep_code}
\alias{keep_code}
\title{keep_code --- Copy code to clipboard or write a specified file.
OUTPUT = Copies code to the clipboard or writes out code from an
object to a specified file.}
\usage{
keep_code(x, file = NULL)
}
\arguments{
\item{x}{Line(s) of code.}

\item{file}{Path to write code. If not specified, code is copied to
the clipboard.}
}
\value{
The code is copied to the clipboard or written to disk.
}
\description{
Simplifies writing and copying code to clipboard by removing the
need to separately create and specify a file connection or to send
code to the clipboard.
}
\examples{
 code <- c("x <- 1", "y <- x + 1")
 keep_code(code) # copies to clipboard by default
}
\author{
Matthew K. Lau
}
