% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{min_graph}
\alias{min_graph}
\title{min_graph --- Minimize a graph to a list of paths.}
\usage{
min_graph(vp = "variable path", vl = "variable lineage")
}
\description{
min_graph --- Minimize a graph to a list of paths.
}
