% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.spine.R
\name{p.spine}
\alias{p.spine}
\title{p.spine --- Parentage spine for a given object.}
\usage{
p.spine(g = "graph", node.id, direction = "in", sep = "_")
}
\arguments{
\item{g}{A graph in matrix format that presents the use and
creation of variables by steps (i.e. complexes of function and
operation calls).}

\item{node.id}{The name of the node of interest.}

\item{direction}{Determines the direction of searching on the graph,
either "in" or "out".}

\item{sep}{Character string used as a separator in the unique IDs of
object nodes.}
}
\value{
A character vector of steps and objects found along the path
    for the object identified by the node ID.
}
\description{
Isolates the line of parentage for an object. This can be either
all steps leading to an object or all steps leading out from an
object.
}
\author{
Matthew K. Lau
}
