% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.graph.R
\name{parse.graph}
\alias{parse.graph}
\title{.parse.graph --- Prases the PROV-JSON formatted output
OUTPUT = A symmetric matrix of provenance entity reltionships.}
\usage{
parse.graph(prov, rm.env = TRUE)
}
\arguments{
\item{prov}{PROV-JSON formatted provenance.}

\item{rm.env}{LOGICAL: should the environment entity be removed?}
}
\value{
A symmetric matrix of provenance entity relationships.
}
\description{
This is an internal function for parsing PROV-JSON formatted 
provenance.
}
\examples{
\dontrun{
data(prov_json)
prov <- read.prov(prov_json)
parse.graph(prov)
}
}
\author{
Matthew K. Lau
}
