% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.info.R
\name{parse.info}
\alias{parse.info}
\title{.parse.info --- Parse node information from PROV-JSON.
OUTPUT = A matrix of node information.}
\usage{
parse.info(prov)
}
\arguments{
\item{prov}{PROV-JSON formatted provenance.}
}
\value{
A symmetric matrix of provenance entity relationships.
}
\description{
This is an internal function for parsing PROV-JSON formatted 
provenance.
}
\examples{
 data(prov_json)
 prov <- jsonlite::fromJSON(prov_json)
 parse.info(prov)
}
\author{
Matthew K. Lau
}
