% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCleaner.R
\name{RCleaner}
\alias{RCleaner}
\title{RClean gadget}
\usage{
RCleaner(data, theme = "united", ...)
}
\arguments{
\item{data}{The data to be used}

\item{theme}{Shiny theme used for output}
}
\description{
interactive data cleaning
}
