% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prov.R
\name{read_prov}
\alias{read_prov}
\title{read_prov --- Read and parse provenance from a JSON file.
OUTPUT = Returns a dataframe containing the provenance.}
\usage{
read_prov(file)
}
\arguments{
\item{file}{JSON formatted object or path to PROV-JSON formatted file.}
}
\value{
A list containing the provenance.
}
\description{
This is an internal function for reading PROV-JSON formatted
provenance.
}
\examples{
 data(prov_json)
 prov <- read_prov(prov_json)
 prov$graph
 prov$info$activity
 prov$info$environment
 prov$info$entity
}
\author{
Matthew K. Lau
}
