% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2chunks.R
\name{rmd2chunks}
\alias{rmd2chunks}
\title{rmd2chunks --- Isolates the code chunks from an R markdown documents.}
\usage{
rmd2chunks(x, comment = "## RMD ##", rm.markdown = FALSE)
}
\arguments{
\item{x}{A character of of code from an R markdown notebook.}

\item{comment}{The string to be used to comment markdown syntax.}

\item{rm.markdown}{LOGICAL: should markdown syntax be removed?}
}
\value{
A character vector consisting of the code chunks isolated from the R markdown notebook code.
}
\description{
To facilitate code cleaning, this function permits users to isolate
the code chunks from an R markdown formatted notebook.
}
\seealso{
write.code
}
\author{
Matthew K. Lau
}
