% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_my_app.R
\name{run_my_app}
\alias{run_my_app}
\title{RClean}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{character string for the directory of this package}

\item{...}{additional options passed to shinyAppDir}
}
\value{
a printed shiny app
}
\description{
Run as a shiny app
}
\examples{
#using example is a complete seperate document, use this
\dontrun{ run_my_app('my_first_app')}

}
