% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RClean.R
\name{run_RClean}
\alias{run_RClean}
\title{RClean}
\usage{
run_RClean(data, xvar, yvar)
}
\arguments{
\item{data}{dataset to use}

\item{xvar}{x variable to plot}

\item{yvar}{y variable to plot}
}
\value{
a shiny gadget output
}
\description{
Run as a shiny gadget
}
\examples{
#using example is a complete seperate document, use this
\dontrun{ run_RClean(mtcars, "hp", "mpg")}


}
