% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_script.R
\docType{data}
\name{simple_script}
\alias{simple_script}
\title{Simple R script.}
\format{A character vector of code lines.}
\usage{
simple_script
}
\description{
Example script that is used in the Rclean examples.
}
\keyword{datasets}
