% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.id.R
\name{var.id}
\alias{var.id}
\title{var.id --- Generate unique identifiers for variables.}
\usage{
var.id(x, sep = "_")
}
\arguments{
\item{x}{A data frame containing "step", "used" by, "defined" by and "var" (i.e. variable) information.}

\item{sep}{Text separator for the identification tags.}
}
\value{
A vector of unique identifiers based on the order of creation in the script.
}
\description{
Uses the order in which variables are created to assign unique identifiers.
}
\author{
Matthew K. Lau
}
