% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.code.R
\name{write.code}
\alias{write.code}
\title{write.code --- Write code to disk.
OUTPUT = Writes out code from an object to a specified file.}
\usage{
write.code(x, file = NULL)
}
\arguments{
\item{x}{Script object to write.}

\item{file}{Path to where you want to write. Default writes to clipboard.}
}
\value{
The script is written to disk, line by line.
}
\description{
This is a simple function for writing code to disk.
}
\author{
Matthew K. Lau
}
