% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean --- Clean code.}
\usage{
clean(script, vars, format = TRUE)
}
\arguments{
\item{script}{File path to a script or a Script object from
CodeDepends.}

\item{vars}{The name of the variable(s) of interest.}

\item{format}{LOGICAL: should the minimized code be re-formatted
following common syntax style guides?}
}
\value{
A character vector containing a minimized script based on
    the given input variable or set of variables.
}
\description{
Reduce the size and complexity of a analytical script through
algorithmic detection of the essential code needed to produce a set
of results.
}
\examples{
script <- system.file(
        "example", 
        "simple_script.R", 
        package = "Rclean")
clean(script)
clean(script, "mat")
clean(script, "tab.12")
clean(script, c("mat", "tab.12", "out")
}
\author{
Matthew K. Lau
}
