% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean --- Clean code.}
\usage{
clean(file, vars, format = TRUE)
}
\arguments{
\item{file}{File path to a script.}

\item{vars}{The name of the variable(s) of interest.}

\item{format}{LOGICAL: should the minimized code be re-formatted
following common syntax?}
}
\value{
A character vector containing a minimized script based on
    the given input variable.
}
\description{
Reduce the size and complexity of a analytical script through
algorithmic detection of the essential code needed to produce a set
of results.
}
\examples{
script <- readScript(
    system.file(
        "example", 
        "simple_script.R", 
        package = "Rclean"))
clean(script)
clean(script, "x")
clean(script, "z")
clean(script, "zz")
}
\author{
Matthew K. Lau
}
