% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanR.R
\name{cleanR}
\alias{cleanR}
\title{cleanR --- Produce "clean" code.
OUTPUT = The essential code needed to produce a result.}
\usage{
cleanR(result = "Name of desired result", file = "Path to script",
  tidy = TRUE)
}
\arguments{
\item{result}{A desired output present in the script.}

\item{file}{Path to an R script.}

\item{tidy}{LOGICAL: should the cleaned script be formatted using syntax best practices?}
}
\value{
Cleaned-up code as a vector of strings ordered by line number.
}
\description{
Produces simplifed, "cleaned" code that is needed to create a result.
}
\author{
Matthew K. Lau
}
