% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_graph.R
\name{code_graph}
\alias{code_graph}
\title{code_graph --- A network diagram of computational dependencies.}
\usage{
code_graph(g)
}
\arguments{
\item{g}{Matrix representation of the graph.}
}
\value{
A network plot.
}
\description{
Plot the dependencies among functions and objects for a given script.
}
\author{
Matthew K. Lau
}
