% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_graph.R
\name{code_graph}
\alias{code_graph}
\title{code_graph --- A network diagram of computational dependencies.}
\usage{
code_graph(src)
}
\arguments{
\item{src}{Script object from CodeDepends..}
}
\value{
A network plot.
}
\description{
Plot the dependencies among functions and objects for a given script.
}
\examples{
simple.script <- readScript(system.file(
    "example", "simple_script.R", package = "Rclean"))
code_graph(simple.script)
}
\author{
Matthew K. Lau
}
