% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libs.R
\name{get_libs}
\alias{get_libs}
\title{get_libs --- Determine the libraries referenced in a given script.}
\usage{
get_libs(file)
}
\arguments{
\item{file}{Path to the script.}
}
\value{
Returns the libraries referenced in the script.
}
\description{
get_libs --- Determine the libraries referenced in a given script.
}
\author{
Matthew K. Lau
}
