% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_script.R
\docType{data}
\name{long_script}
\alias{long_script}
\title{Long R script.}
\format{A vector of code lines.
\describe{
  \item{simple_script}{The script.}
}}
\usage{
long_script
}
\description{
Example script that is used in the Rclean examples.
}
\keyword{datasets}
