% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCleaner.R
\name{RCleaner}
\alias{RCleaner}
\title{RClean gadget}
\usage{
RCleaner(Data, theme = "united", ...)
}
\arguments{
\item{theme}{Shiny theme used for output (default is united)}

\item{...}{additional arguments not specified_gh}

\item{data}{The data to be used - must be of type DF, tibble or matrix with 2 or more columns}
}
\description{
Interactive data cleaning.  This gadget initializes a DF, matrix or tibble in a browser window for data cleaning.
}
