% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCleaner.R
\name{RCleaner}
\alias{RCleaner}
\title{RClean gadget}
\usage{
RCleaner(data, theme = "united", ...)
}
\arguments{
\item{data}{The data to be used - must be of type DF, tibble or matrix with 2 or more columns}

\item{theme}{Shiny theme used for output}

\item{...}{additional arguments not specified}
}
\description{
interactive data cleaning
}
