% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_script.R
\docType{data}
\name{simple_script}
\alias{simple_script}
\title{Simple R script.}
\format{A vector of code lines:
\describe{
  \item{simple_script}{The script.}
  ... 
}}
\usage{
simple_script
}
\description{
Example script that is used in the Rclean examples.
}
\keyword{datasets}
